==> 02. Scraping mountain pass coordinates.ipynb <==
# Obtaining the coordinates for each mountain pass in Spain

Having the exact coordinates for every port in Spain would allow us to determine exactly which routes pass through them, greatly enhancing the capabilities of our dataset. But where could we find those coordinates?

The website **Altimetrias.net** has a handy map with the coordinates for 757 ports, 745 of them being in continental Spain. In the first part of this notebook I will try to extract their coords using web scraping, but since our database has ~1700 ports the quantity of missing information at this stage is not trivial.

## Extracting the coordinates for *Altimetrias.net* 757 ports

As I already mentioned I'm going to use web scraping at this stage. Let's begin by importing our libraries.

import requests
from bs4 import BeautifulSoup

#The map url.

url = 'https://www.altimetrias.net/mapas/default.asp'

#Parsing the content to make it accessible.

html = requests.get(url).text
soup = BeautifulSoup(html, 'html.parser')

## Accessing the data

The port names, coordinates and urls are contained in *script* objects. We can simply isolate and iterate through them in order to extract the needed data.

import re

#We find our first port at the index value 3, the last one being at the index position 760.

puerto = soup.find_all('script')[3]

#Visualizing the data structure within the element.

puerto.string

#Accessing the latitude.

re.search("\'(.*?)\', \'", puerto.string)[1]

#Accessing the longitude.

re.search(", \'(.*?)\'],", puerto.string)[1]

#Accessing the url.

re.search('href="(.*?)"><b', puerto.string)[1]

## Using a loop to extract the parsed data

Now that we know the exact location and format of the data it's simply a matter of creating a loop that will iterate through every *script* object and return a dictionary. We will later use those dictionaries to create a dataframe.

dict_list = []

for i in range(3,760):
    puerto = soup.find_all('script')[i]
    dict_puerto = {'url': re.search('href="(.*?)"><b', puerto.string)[1],
                    'lat': re.search("\'(.*?)\', \'", puerto.string)[1],
                    'long': re.search(", \'(.*?)\'],", puerto.string)[1]}
    dict_list.append(dict_puerto)

#Let's check how many ports we obtained.

len(dict_list)

#Testing the dictionary list.

dict_list[0]

## Creating a dataframe

Now that we have obtained all the usaful data we will package it into a dataframe.

import pandas as pd

df = pd.DataFrame(dict_list)

df.head()

df.info()

#Saving our dataframe.

df.to_csv('df_coords_puertos.csv', index=False)

# Obtaining the coordinates for the remaining ports

Let's load out our ports dataframe to find out which ones have missing coordinates.

df1 = pd.read_csv('master_df_puertos.csv')

df1.head()

df1.info()

#Some ports are too small to be of any value, so we will be dropping them and creating a new df.

df2 = df1[df1['desnivel'] >= 200]

df2.info()

## Obtaining the missing ports

First of all we will merge the two dataframes to visualize which ports are lacking coordinates.

df3 = pd.merge(left=df2, right=df, how='left', left_on='url', right_on='url')

df3.info()

df3.head()

#Creating a dataframe with all ports that are missing their coordinates.

df4 = df3[df3['lat'].isnull()]

df4.info()

df4.head()

#Exporting and importing to tweak some port names manually.

df4.to_csv('df4.csv', index=False)

df4 = pd.read_csv('df4.csv')

## Requesting the missing coordinates

For this purpose I will be using a library called **Geopy**.

from geopy.geocoders import Nominatim

#Testing the function.

geolocator = Nominatim(user_agent="Firefox")
location = geolocator.geocode("Pico Veleta, Granada, España")
location

#Obtaining the longitude:

location[1][0]

#Latitude:

location[1][1]

#Let's create a loop that iterates through our dataframe and appends coordinates whenever possible.

for i in range(len(df4)):
    try:
        geolocator = Nominatim(user_agent="Firefox")
        puerto = df4['puerto'].iloc[i]
        provincia = df4['provincia'].iloc[i]
        buscador = puerto + ', ' + provincia + ', ' + 'España'
        location = geolocator.geocode(buscador)
        df4['lat'].iloc[i] = location[1][0]
        df4['long'].iloc[i] = location[1][1]
    except:
        pass

#We obtained the coordinates for 765 additional ports. The next step will be creating two dataframes, one with all ports that 
#have coords and one without them.

df4.info()

## Creating the final dataframes

#First, the dataframe with missing coordinates.

df5 = df4[df4['lat'].isnull()]

df5.info()

#Saving our dataframe of missing coordinates.

df5.to_csv('df_missing_routes.csv', index=False)

#And second, the dataframe with coords.

df6 = df4.dropna(how='any')

df6.info()

#Merging the final dataframes.

df7 = df3.dropna(how='any')

df7.info()

#Appending the dataframe.

df8 = df7.append(df6, ignore_index=True)

df8.head()

df8.info()

#Exporting our dataframe:

df8.to_csv('master_df_puertos_coords.csv', index=False)

# Conclusions

While we managed to obtain the coordinates for 1109 ports, 359 still have missing data. Most of them are quite obscure, but if we are aiming for completeness those values will have to be extracted manually.

Thankfully this won't really hurt our results since we have the geographical data for the most important ports and 1167 is plenty of data to work with.

**<div align="right">Ironhack DA PT 2021</div>**
    
**<div align="right">Xavier Esteban</div>**
