==> 03. Routes scraping (I).ipynb <==
# Scraping the best routes for each destination from Wikiloc.

Now that we have all mountain passes (both its name and coordinates) we will begin by scraping all road bike routes on **Wikiloc** that match the name of a mountain pass. Our tool of choice will be the web scraping engine **Selenium**.

#Importing libraries.

import pandas as pd
import requests
from selenium import webdriver
from selenium.webdriver.common.keys import Keys
from bs4 import BeautifulSoup
import time

#Initializing our webdriver (Chrome).

driver = webdriver.Chrome()

#Logging into Wikiloc.

def login():
    time.sleep(1)
    driver.get('https://es.wikiloc.com/wikiloc/start.do')
    time.sleep(1)
    login_button = driver.find_element_by_class_name('form-control')
    login_button.click()
    login_button.send_keys('username')
    password_button = driver.find_element_by_xpath('//*[@id="password"]')
    password_button.click()
    password_button.send_keys('password')

#At this point we MUST solve the captcha manually. Once solved we can proceed.

#Navigating to the search section with our "road bike" filter activated.

def filtro():
    search_bar = driver.find_element_by_xpath('/html/body/header/nav/div/div[1]/div/div')
    search_bar.click()
    time.sleep(1)
    search_bar_open = driver.find_element_by_xpath('/html/body/header/nav/div/div[1]/div/div[1]/input')
    search_bar_open.send_keys('bracons')
    time.sleep(1)
    bracons = driver.find_element_by_xpath('/html/body/header/nav/div/div[1]/div/div[2]/div/ul/div')
    bracons.click()
    time.sleep(1)
    actividad = driver.find_element_by_xpath('/html/body/div[1]/div[1]/div[1]/div/div[1]/button')
    actividad.click()
    time.sleep(1)
    selector_actividad = driver.find_element_by_xpath('/html/body/div[1]/div[1]/div[1]/div/div[1]/div/div/div/div[1]/div[1]/div/div[2]/div/div[2]/div')
    selector_actividad.click()
    time.sleep(1)
    aplicar_filtro = driver.find_element_by_xpath('/html/body/div[1]/div[1]/div[1]/div/div[1]/div/div/div/div[2]/button')
    aplicar_filtro.click() #Selecting the "road bike" filter. Now we're all set to scrape any destination.

#This function scrapes all relevant data from the found routes and returns it as a list of dictionaries.

def scraper_ruta(puerto):
    listado_dict = []
    element = driver.find_elements_by_class_name('trail-list__wrapper')
    for i in range(len(element)):
        try:
            element[i].click()
            time.sleep(0.3)
            nombre_ruta = driver.find_elements_by_xpath('//*[@id="front"]/div/div/h1')[0].text
            url = driver.current_url
            trailrank = driver.find_elements_by_xpath('//*[@id="front"]/div/div/a')[0].text
            distancia = driver.find_elements_by_xpath('//*[@id="trail-data"]/div[1]/div[1]/a/span[2]')[0].text
            desnivel = driver.find_elements_by_xpath('//*[@id="trail-data"]/div[1]/div[3]/a/span[2]')[0].text
            dificultad = driver.find_elements_by_xpath('//*[@id="trail-data"]/div[1]/div[7]/a/span[2]')[0].text
            fotos = driver.find_elements_by_class_name('trail-photo')
            photo1 = fotos[0].get_attribute('href')
            photo2 = fotos[1].get_attribute('href')
            photo3 = fotos[2].get_attribute('href')
            driver.back()
            time.sleep(0.3)
            element = driver.find_elements_by_class_name('trail-list__wrapper')
            dict_ruta = {'ubicacion': puerto, 'nombre': nombre_ruta, 'trailrank': trailrank,'distancia': distancia, 'desnivel': desnivel,
                         'dificultad': dificultad, 'url': url, 'photo1': photo1, 'photo2': photo2, 'photo3': photo3}
            listado_dict.append(dict_ruta)
        except:
            pass
            driver.back()
            time.sleep(0.5)
            element = driver.find_elements_by_class_name('trail-list__wrapper')
    return listado_dict

#Defining a function that will search routes for the requested destination.

def buscar_ubicacion(ubicacion):
    time.sleep(0.3)
    driver.find_elements_by_xpath('/html/body/header/nav/div/div[1]/div/div/input')[0].click()
    time.sleep(0.3)
    driver.find_elements_by_xpath('/html/body/header/nav/div/div[1]/div/div[1]/input')[0].send_keys(ubicacion)
    time.sleep(0.5)
    driver.find_element_by_xpath('/html/body/header/nav/div/div[1]/div/div[2]/div/ul/div').click()
    time.sleep(0.5)

df_puertos = pd.read_csv('df_puertos_final.csv')

df_puertos.head()

#Creating a list containing all mountain pass names.

lista_ubicaciones = df_puertos['puerto']
lista_ubicaciones = lista_ubicaciones.to_list()

len(lista_ubicaciones)

## Final function

#This function incorporates all previous functions, iterating through each destination and returning a dataframe of routes.

def scraper_puertos(puertoncios):
    start = time.time()
    lista = []
    for i in puertoncios:
        try:
            buscar_ubicacion(i)
            lista.append(scraper_ruta(i))
        except:
            pass
    df = pd.DataFrame(lista[0])
    for i in lista[1:]:
        df = df.append(i, ignore_index=True)
    stop = time.time() 
    duration = (stop - start) / 60
    print('Minutos:', duration)
    return df

#Benchmarking our function.

df_test = scraper_puertos(lista_ubicaciones[])

#Creating a dataframe of routes using the function.

df_test.to_csv('df_test.csv', index=False)

df_test.info()

## Cleaning our dataframe

Now that we've obtained a dataframe of our routes, some data cleaning is needed.

df1 = pd.read_csv('df_0_50.csv') #Loading all obtained csv files as not to re-run the code.
df2 = pd.read_csv('df_50_100.csv')
df3 = pd.read_csv('df_100_150.csv')
df4 = pd.read_csv('df_150_end.csv')

df = pd.concat([df1, df2, df3, df4], ignore_index=True) #Uniting them in a single DF.

df.drop('Unnamed: 0', inplace=True, axis=1) #Dropping a trash column.

df.head()

df.to_csv('routes.csv')

#Assigning a numerical score to the route difficulty.

df.loc[df['dificultad'] == 'Fácil', 'dificultad'] = 1
df.loc[df['dificultad'] == 'Moderado', 'dificultad'] = 2
df.loc[df['dificultad'] == 'Difícil', 'dificultad'] = 3
df.loc[df['dificultad'] == 'Muy difícil', 'dificultad'] = 4
df.loc[df['dificultad'] == 'Sólo expertos', 'dificultad'] = 5

#We see that we've got 526 locations, 49 less than in our 'altimetrias' dataset. This means that 49 locations didn't have
#routes matching our quality criteria.

df['ubicacion'].value_counts() 

df.sort_values(by=['trailrank'], ascending=False).head(50)

**<div align="right">Ironhack DA PT 2021</div>**
    
**<div align="right">Xavier Esteban</div>**
