==> 05. Routes scraping (II) and GPX parsing.ipynb <==
# Downloading routes that pass through a port's coordinates

Until now our route search was quite imprecise since it was based on searching for a port name on *Wikiloc*. Now that we have the exact coordinates for most ports it's quite possible to use the site's *'search routes that pass through here'* option and find routes that pass **exactly** through each port. 

Let's get to it.

## Preparing our setup

For this purpose we will need to load our usual libraries, import a few functions and load the dataframe of port coordinates.

import pandas as pd
import requests
from selenium import webdriver
from selenium.webdriver.common.keys import Keys
from bs4 import BeautifulSoup
import time
import folium
import re
import pathlib
import os
import gpxpy
import gpxpy.gpx
import time
from pathlib import Path
import os

df = pd.read_csv('master_df_puertos_coords.csv')

driver = webdriver.Chrome()
driver.get('https://es.wikiloc.com/wikiloc/start.do')

#Navigating to the search section with our "road bike" filter activated.

def filtro():
    search_bar = driver.find_element_by_xpath('/html/body/header/nav/div/div[1]/div/div')
    search_bar.click()
    time.sleep(1)
    search_bar_open = driver.find_element_by_xpath('/html/body/header/nav/div/div[1]/div/div[1]/input')
    search_bar_open.send_keys('bracons')
    time.sleep(1)
    bracons = driver.find_element_by_xpath('/html/body/header/nav/div/div[1]/div/div[2]/div/ul/div')
    bracons.click()
    time.sleep(1)
    actividad = driver.find_element_by_xpath('/html/body/div[1]/div[1]/div[1]/div/div[1]/button')
    actividad.click()
    time.sleep(1)
    selector_actividad = driver.find_element_by_xpath('/html/body/div[1]/div[1]/div[1]/div/div[1]/div/div/div/div[1]/div[1]/div/div[2]/div/div[2]/div')
    selector_actividad.click()
    time.sleep(1)
    aplicar_filtro = driver.find_element_by_xpath('/html/body/div[1]/div[1]/div[1]/div/div[1]/div/div/div/div[2]/button')
    aplicar_filtro.click() #Selecting the "road bike" filter. Now we're all set to scrape any destination.

filtro()

Our previous scrapper function only took into account the search location's name, but this time we will be including coordinates to get a far more accurate result.

# Using our function to scrape all routes

In this section we created a loop that makes use of **Wikiloc**'s function "search routes through this point" to pass our port coordinates and find routes that reach every mountain pass.

#Creating a reduced dataframe to test our function.

df_10 = df.head(10)

#Defining and testing our loop.

start = time.time() #Starting a timer.

lista_rutas = []

for i in range(len(df_10)):
    try:
        time.sleep(0.5)
        url = 'https://es.wikiloc.com/wikiloc/map.do?sw=41.55952720430883%2C2.3658370971679688&ne=41.58575748995264%2C2.415275573730469&act=29&zdp=(' + str(df['long'].iloc[i]) + '%2C' + str(df['lat'].iloc[i]) + '%2C357)&page=1'
        driver.get(url)
        time.sleep(0.5)
        listado_dict = []
        element = driver.find_elements_by_class_name('trail-list__wrapper')
        for n in range(len(element)):
            try:
                element[n].click()
                time.sleep(0.3)
                nombre_ruta = driver.find_elements_by_xpath('//*[@id="front"]/div/div/h1')[0].text
                url = driver.current_url
                trailrank = driver.find_elements_by_xpath('//*[@id="front"]/div/div/a')[0].text
                distancia = driver.find_elements_by_xpath('//*[@id="trail-data"]/div[1]/div[1]/a/span[2]')[0].text
                desnivel = driver.find_elements_by_xpath('//*[@id="trail-data"]/div[1]/div[3]/a/span[2]')[0].text
                dificultad = driver.find_elements_by_xpath('//*[@id="trail-data"]/div[1]/div[7]/a/span[2]')[0].text
                fotos = driver.find_elements_by_class_name('trail-photo')
                photo1 = fotos[0].get_attribute('href')
                photo2 = fotos[1].get_attribute('href')
                photo3 = fotos[2].get_attribute('href')
                driver.back()
                time.sleep(0.3)
                element = driver.find_elements_by_class_name('trail-list__wrapper')
                dict_ruta = {'ubicacion': df_10['puerto'].iloc[i], 'nombre': nombre_ruta, 'trailrank': trailrank,'distancia': distancia, 'desnivel': desnivel,
                             'dificultad': dificultad, 'url': url, 'photo1': photo1, 'photo2': photo2, 'photo3': photo3}
                lista_rutas.append(dict_ruta)
            except:
                pass
                driver.back()
                time.sleep(0.5)
                element = driver.find_elements_by_class_name('trail-list__wrapper')
    except:
        time.sleep(2)
        pass
    
stop = time.time() 
duration = (stop - start) / 60
print('Minutes:', duration)

#Creating a dataframe out of the list of routes.

df_rutas = pd.DataFrame(lista_rutas)

df_rutas.info()

We can now scrape all routes. It's going to take a while.

start = time.time() #Starting a timer.

lista_rutas = []

for i in range(len(df)):
    try:
        time.sleep(0.5)
        url = 'https://es.wikiloc.com/wikiloc/map.do?sw=41.55952720430883%2C2.3658370971679688&ne=41.58575748995264%2C2.415275573730469&act=29&zdp=(' + str(df['long'].iloc[i]) + '%2C' + str(df['lat'].iloc[i]) + '%2C357)&page=1'
        driver.get(url)
        time.sleep(0.5)
        listado_dict = []
        element = driver.find_elements_by_class_name('trail-list__wrapper')
        for n in range(len(element)):
            try:
                element[n].click()
                time.sleep(0.3)
                nombre_ruta = driver.find_elements_by_xpath('//*[@id="front"]/div/div/h1')[0].text
                url = driver.current_url
                trailrank = driver.find_elements_by_xpath('//*[@id="front"]/div/div/a')[0].text
                distancia = driver.find_elements_by_xpath('//*[@id="trail-data"]/div[1]/div[1]/a/span[2]')[0].text
                desnivel = driver.find_elements_by_xpath('//*[@id="trail-data"]/div[1]/div[3]/a/span[2]')[0].text
                dificultad = driver.find_elements_by_xpath('//*[@id="trail-data"]/div[1]/div[7]/a/span[2]')[0].text
                fotos = driver.find_elements_by_class_name('trail-photo')
                photo1 = fotos[0].get_attribute('href')
                photo2 = fotos[1].get_attribute('href')
                photo3 = fotos[2].get_attribute('href')
                driver.back()
                time.sleep(0.3)
                element = driver.find_elements_by_class_name('trail-list__wrapper')
                dict_ruta = {'ubicacion': df['puerto'].iloc[i], 'nombre': nombre_ruta, 'trailrank': trailrank,'distancia': distancia, 'desnivel': desnivel,
                             'dificultad': dificultad, 'url': url, 'photo1': photo1, 'photo2': photo2, 'photo3': photo3}
                lista_rutas.append(dict_ruta)
            except:
                pass
                driver.back()
                time.sleep(0.5)
                element = driver.find_elements_by_class_name('trail-list__wrapper')
    except:
        time.sleep(2)
        pass
    
stop = time.time() 
duration = (stop - start) / 60
print('Minutes:', duration)

df_rutas = pd.DataFrame(lista_rutas)
df_rutas.info()

#Saving the resulting dataframe.

df_rutas.to_csv('df_extra_routes.csv', index=False)

#Adding favs function.

t = 0.4 #This will set the wait between actions. If you have a good connection you can keep it between 0.3-0.5. 
        #Some trial and error might be necessary.

def fav(url_list):
    start = time.time() #Starting a counter to time our code.
    for i in url_list: #url_list will be the list of the route's urls (max 1000 and no duplicates, VERY important).
        try:
            driver.get(i) #Accessing the url.
            time.sleep(t) 
            driver.find_element_by_xpath('//*[@id="container"]/a').click() #Clicking on the 'Add to favorites' item.
            time.sleep(t)
            driver.find_element_by_xpath('//*[@id="container"]/div/div/div/div[3]/div[1]').click() #Marking the route as fav.
            time.sleep(t)
        except:
            time.sleep(1) 
            pass
    stop = time.time() #Stopping our timer.
    duration = (stop - start) / 60 #Calculating the elapsed minutes.
    print(len(url_list), 'favs added in', duration, 'minutes.')

url_list = df_rutas['url'].tolist()

fav(url_list[:1000])

# Checking for missing routes

First of all we must delete duplicate routes.

df = pd.read_csv('df_extra_routes.csv')

df.sort_values(by='trailrank', ascending=False, inplace=True)

df.drop_duplicates(subset=['url'], keep='first', inplace=True)

df.head()

df.info()

Performing the necessary cleaning.

df['distancia'] = df['distancia'].str.replace(" km", "")
df['distancia'] = df['distancia'].str.replace(",", ".")
df['desnivel'] = df['desnivel'].str.replace(" m", "")
df['desnivel'] = df['desnivel'].str.replace(".", "")

#Adding the alpha name column.

def stripper(name):
    return re.sub(r'\W+', '', name)[:20]

df['alpha_name'] = df['nombre'].apply(stripper)

#Renaming every gpx file with the new shortened and stripped name, matching our new df column.

for path in pathlib.Path("/gpx").iterdir(): #Using iterdir to iterate through every file in our gpx folder. 
    try:
        if path.is_file():
            old_name = path.stem
            old_extension = path.suffix
            directory = path.parent
            strip = re.sub(r'\W+', '', old_name) #Stripping the name using the same regex as before.
            new_name = strip[:20] + old_extension #Only keeping the first 20 characters.
            path.rename(pathlib.Path(directory, new_name)) #Renaming the file.
    except:
        pass

#Creating a list of our shortened alphanumeric names.

alpha_list = df['alpha_name'].tolist()
len(alpha_list)

#Now that our dataframe and gpx files match we can easily check for missing entries.

missing_routes = [] #This list will store our missing filenames.
ok_routes = [] #Successful downloads.

for i in alpha_list:
    gpx_path = 'gpx/' + i + '.gpx'
    isExist = os.path.exists(gpx_path) #Checking if the file with the given filename exists.
    if isExist == True:
        ok_routes.append(i)
        pass
    else:
        missing_routes.append(i)

len(missing_routes)

#Creating a list of missing routes url.

missing_list = []

for i in range(1138):
    if df['alpha_name'].iloc[i] in missing_routes:
        missing_list.append(df['url'].iloc[i])

len(missing_list)

fav(missing_list)

## Parsing all gpx files

Now that we have all files it's time to parse them so that we can extract valuable data and further analyze them. For this purpose we will be using our previously created parser function.

This function takes a given gpx file and returns a dictionary of the most important values, thus making it easy to convert a list of such dictionaries into a dataframe.

def parser(file):
    try:
        gpx_file = open(file, 'r', encoding='utf-8') 
        gpx = gpxpy.parse(gpx_file) 
        coords = []
        alt = []
        for track in gpx.tracks:
            for segment in track.segments:        
                for point in segment.points:
                    coords.append(tuple([point.latitude, point.longitude]))
                    alt.append(point.elevation)
        parsed_file = {'name': track.name, 
                       'coords': [coords], 
                       'alt': [alt], 
                       'distance': track.length_3d()/1000, 
                       'climb': int(gpx.get_uphill_downhill()[0]),
                       'min_alt': int(gpx.get_elevation_extremes()[0]),
                       'max_alt': int(gpx.get_elevation_extremes()[1]),
                       'start': coords[0][0],
                       's_lo': coords[0][1],
                       'f_la': coords[-1][0],
                       'f_lo': coords[-1][1],}
        return parsed_file
    except:
        pass

start = time.time()

dict_list = []

directory = 'gpx' #Our very original gpx folder name.
 
files = Path(directory).glob('*')
for file in files:
    dict_list.append(parser(file))
    
stop = time.time() 
duration = (stop - start) / 60
print('Minutes:', duration)

len(dict_list)

df9 = pd.DataFrame(dict_list)

df9.to_csv('gpx_9.csv', index=False)

df1 =pd.read_csv('gpx_1.csv')
df2 =pd.read_csv('gpx_2.csv')
df3 =pd.read_csv('gpx_3.csv')
df4 =pd.read_csv('gpx_4.csv')
df5 =pd.read_csv('gpx_5.csv')
df6 =pd.read_csv('gpx_6.csv')
df7 =pd.read_csv('gpx_7.csv')
df8 =pd.read_csv('gpx_8.csv')
df9 =pd.read_csv('gpx_9.csv')

df_parsed = pd.concat([df1, df2, df3, df4, df5, df8, df7, df8, df9])

df_parsed.drop_duplicates('name', keep='first', inplace=True)

df_parsed.info()

df_parsed.to_csv('master_parsed_all.csv', index=False)

df_parsed.head()

## 3.1 Merging the parsed gpx dataframe with the routes dataframe

#Loading our routes dataframe.

df_rutas = pd.read_csv('master_rutas_final.csv')

#Merging the parsed gpx files with the corresponding routes to create a final dataframe.

df_merge = pd.merge(df_rutas, df_parsed, how='inner', left_on=['nombre'], right_on=['name'])

#We now have 9171 parsed routes, but some of them will be outside Spain and will have to be discarded.

df_merge.info()

**<div align="right">Ironhack DA PT 2021</div>**
    
**<div align="right">Xavier Esteban</div>**
