==> 06. Final dataset creation.ipynb <==
# Final dataframe creation

After all  transformations have been made we want to have just 3 dataframes that contain all useful information:

**1. Routes:** the best cycling routes to climb each mountain pass.

**2. Towns:** all towns with 3 or more routes that pass near them.

**3. Ports:** all geolocated mountain ports.

## Routes: general data cleaning

import pandas as pd
import time
import haversine as hs

routes = pd.read_csv('routes_1607_819.csv')

routes.info()

#Adding _id and municipalities_ids columns:

routes['municipality'] = None

routes.info()

#Renaming the columns.

routes.rename(columns = {'nombre': 'name', 'provincia': 'province', 'distancia': 'distance', 'desnivel': 'gradient', 'municipios': 'municipalities_ids', 'puertos': 'mountain_passes_ids'}, inplace = True)

#Re-ordering the columns.

routes = routes[['ID', 'name', 'ccaa', 'province', 'start', 'midpoint', 'trailrank', 'distance', 'gradient', 'min_alt', 'max_alt', 'municipality', 'mountain_passes_ids', 'municipalities_ids', 'coords', 'alt']]

#Dropping coords and alt since they can't be loaded into MongoDB.

routes.drop(['coords', 'alt'], axis=1, inplace=True)

routes.head()

## Routes: converting coordinates for MongoDB

**MongoDB** uses a different formatting for latitude/longitude, so we will have to modify our dataset accordingly.

#Defining our function.

def route_converter(df):
    
    """
    Input : dataframe of routes.
    
    Output: dataframe with switched coordinates (lat/long -> long/lat) and substitutes all '(' with '['.
    
    """
    
    for i in range(len(df)):
        df['start'].iloc[i] = '[' + str(eval(df['start'].iloc[i])[1]) + ',' + str(eval(df['start'].iloc[i])[0]) + ']'
        df['midpoint'].iloc[i] = '[' + str(eval(df['midpoint'].iloc[i])[1]) + ',' + str(eval(df['midpoint'].iloc[i])[0]) + ']'
        
    return df

#Converting coords.

routes = route_converter(routes)

#Let's save this dataframe.

routes.to_csv('routes_1807_819i.csv', index=False)

#We also need this dataframe with its coordinates inverted to use it on MongoDB. For this purpose we will be using a function:

def converter(df):
    for i in range(len(df)):
        df['start'].iloc[i] = '[' + str(eval(df['start'].iloc[i])[1]) + ',' + str(eval(df['start'].iloc[i])[0]) + ']'
        df['midpoint'].iloc[i] = '[' + str(eval(df['midpoint'].iloc[i])[1]) + ',' + str(eval(df['midpoint'].iloc[i])[0]) + ']'


    for i in range(len(df)):
        newc = eval(df['coords'].iloc[i])
        c_list = []
        for n in newc:
            c_list.append([n[1],n[0]])
        df['coords'].iloc[i] = c_list
        
    return df

routes = converter(routes)

#Finally, exporting our dataframe ready to use on MongoDB.

routes.to_csv('routes_1607_819_DB.csv', index=False)

## Routes: matching routes with nearby towns

Now that we have all 819 routes it's time to find which towns are near each one. Once this is done we can finally find which towns have 3 or more routes near them, thus making them ideal cycling destinations.

#Let's import our municipalities dataframe.

towns = pd.read_csv('compartir/municipios y rutas.csv')

#This function takes two dataframes (towns and routes) and looks for towns 1.5Km or less from any point of our routes.

def nearby_routes(routes, towns):
    dict_list = []
    for g in range(len(routes)):
        new_c = eval(routes['coords'].iloc[g])
        route_list = []
        for i in new_c[0::60]:
            for n in range(len(towns)):
                try:
                    if hs.haversine((i), eval(towns['coords'].iloc[n])) - towns['radius'].iloc[n] < 1.5:
                        if towns['ID'].iloc[n] not in route_list:
                            route_list.append(towns['ID'].iloc[n])
                        else:
                            pass
                except:
                    pass
        dict_routes = {'route': routes['ID'].iloc[g], 'town': route_list}
        dict_list.append(dict_routes)
    return dict_list

#Using the function on our dataframes. This block of code performs all necessary transformations on it.

start = time.time()

test1 = pd.DataFrame(nearby_routes(routes, towns))

df_exploded = test1.explode('town')

town_list = towns['ID'].tolist()

#Using a simple loop to generate a dictionary of nearby routes for every town, and adding that dictionary to a list.

dict_list = []

for i in town_list:
    try: 
        lista_rutas = []
        for n in range(len(df_exploded)):
            if df_exploded['town'].iloc[n] == i:
                lista_rutas.append(df_exploded['route'].iloc[n])
            else:
                pass
        dict_ruta = {'municipio': i, 'rutas': lista_rutas, 'numero_rutas': len(lista_rutas)}
        dict_list.append(dict_ruta)
    except:
        pass
    
#Creating a dataframe out of the dict list.

df_dict = pd.DataFrame.from_dict(dict_list)


final_list  = df_dict['rutas'].tolist()
num_routes = df_dict['numero_rutas'].tolist()
towns['rutas'] = final_list
towns['num_rutas'] = num_routes


stop = time.time() 
duration = (stop - start) / 60
print('Minutes:', duration)

towns.to_csv('towns_1707.csv', index=False)

towns.head(50)

# Towns: general data cleaning

Cleaning and renaming columns on our *towns* dataset. We will also need to convert it for its use with **MongoDB**, thankfully we have a function for that.

towns = pd.read_csv('towns_1707.csv')

towns.info()

#Renaming the columns.

towns.rename(columns = {'provincia': 'province', 'municipio': 'municipality', 'poblacion': 'municipality_inhabitants', 'superficie': 'geographic_area', 'altitud': 'alt', 'num_rutas': 'routes_number', 'rutas': 'routes_ids'}, inplace = True)

#Adding a column to store all mountain passes that can be accessed from each town.

towns['mountain_passes_ids'] = None

#Adding the ccaa column, it will be populated in a few minutes.

towns['ccaa'] = None

#Reordering the columns.

towns = towns[['ID', 'municipality', 'ccaa', 'province', 'municipality_inhabitants', 'geographic_area', 'radius', 'routes_number', 'routes_ids', 'mountain_passes_ids', 'coords']]

towns.info()

## Towns: adding 'ccaa' to each town

My first approach was to use **Geopy** to obtain the **CCAA** of every town, but the process encountered many errors and the library wasn't fully reliable for this purpose. I instead chose to use a simple series of loops to assign the correct **CCAA** to each province. This process, while labor-intensive at first, allows me to re-use this loop on any given dataframe simply by changing some column names.

#Adding the CCAA is easily done by using a simple loop.

for i in range(len(towns)):
    if towns['province'].iloc[i] in ['Álava', 'Bizkaia', 'Gipuzkoa']:
        towns['ccaa'].iloc[i] = 'País Vasco'
    else:
        pass
    
for i in range(len(towns)):
    if towns['province'].iloc[i] in ['Almería', 'Cádiz', 'Córdoba', 'Granada', 'Huelva', 'Jaén', 'Málaga', 'Sevilla']:
        towns['ccaa'].iloc[i] = 'Andalucía'
    else:
        pass
    
for i in range(len(towns)):
    if towns['province'].iloc[i] in ['Huesca', 'Teruel', 'Zaragoza']:
        towns['ccaa'].iloc[i] = 'Aragón'
    else:
        pass
    
for i in range(len(towns)):
    if towns['province'].iloc[i] in ['Asturias']:
        towns['ccaa'].iloc[i] = 'Asturias'
    else:
        pass
    
for i in range(len(towns)):
    if towns['province'].iloc[i] in ['Illes Balears']:
        towns['ccaa'].iloc[i] = 'Illes Balears'
    else:
        pass
    
for i in range(len(towns)):
    if towns['province'].iloc[i] in ['Las Palmas', 'Santa Cruz de Tenerife']:
        towns['ccaa'].iloc[i] = 'Canarias'
    else:
        pass
    
for i in range(len(towns)):
    if towns['province'].iloc[i] in ['Cantabria']:
        towns['ccaa'].iloc[i] = 'Cantabria'
    else:
        pass
    
for i in range(len(towns)):
    if towns['province'].iloc[i] in ['Albacete', 'Ciudad Real', 'Cuenca', 'Guadalajara', 'Toledo']:
        towns['ccaa'].iloc[i] = 'Castilla-La Mancha'
    else:
        pass
    
for i in range(len(towns)):
    if towns['province'].iloc[i] in ['Barcelona', 'Girona', 'Lleida', 'Tarragona']:
        towns['ccaa'].iloc[i] = 'Cataluña'
    else:
        pass
    
for i in range(len(towns)):
    if towns['province'].iloc[i] in ['Badajoz', 'Cáceres']:
        towns['ccaa'].iloc[i] = 'Extremadura'
    else:
        pass
    
for i in range(len(towns)):
    if towns['province'].iloc[i] in ['A Coruña', 'Luga', 'Ourense', 'Pontevedra']:
        towns['ccaa'].iloc[i] = 'Galicia'
    else:
        pass
    
for i in range(len(towns)):
    if towns['province'].iloc[i] in ['Alicante', 'Castellón', 'Valencia']:
        towns['ccaa'].iloc[i] = 'Comunitat Valenciana'
    else:
        pass
    
for i in range(len(towns)):
    if towns['province'].iloc[i] in ['Madrid']:
        towns['ccaa'].iloc[i] = 'Comunidad de Madrid'
    else:
        pass
for i in range(len(towns)):
    if towns['province'].iloc[i] in ['Murcia']:
        towns['ccaa'].iloc[i] = 'Región de Murcia'
    else:
        pass
    
for i in range(len(towns)):
    if towns['province'].iloc[i] in ['Navarra']:
        towns['ccaa'].iloc[i] = 'Navarra'
    else:
        pass
    
for i in range(len(towns)):
    if towns['province'].iloc[i] in ['La Rioja']:
        towns['ccaa'].iloc[i] = 'La Rioja'
    else:
        pass

#Checking the values. Everything seems to be in order.

towns['ccaa'].value_counts()

## Towns: keeping towns with 3 or more routes

# Finally, let's filter out towns with less than 3 routes.

towns = towns[towns['routes_number'] >= 3]

towns.info()

## Towns:  filtering out towns that share the same routes

In the event of several towns sharing the same routes we want to keep the bigger one, since it will have the most services and accomodations. To achieve this we will have to search for duplicates on the *routes_ids* column, but since they are stored in non-sorted lists we will have to order them first.

#Creating a loop that evaluates each route_id string as a list and orders it.

for i in range(len(towns)):
    ordered = eval(towns['routes_ids'].iloc[i])
    ordered.sort()#Evaluating the string as a list and sorting it.
    towns['routes_ids'].iloc[i] = str(ordered) #Replacing the non-ordered list with the new one (as a string).

#All route ids have been neatly ordered.

towns.head(1)

#Now we can simply call drop_duplicates on that column to get rid of duplicate routes, but before that we will order our
#dataframe by town size so that we actually keep the bigger town.

towns = towns.sort_values('municipality_inhabitants', ascending=False)
towns = towns.drop_duplicates('routes_ids', keep='first')

#We're finally left with 841 towns.

towns.info()

#Let's save this dataset before re-arranging its coordinates. The final 'n' marks it as non-inverted.

towns.to_csv('towns_1807_841n.csv', index=False)

## Towns: converting coordinates

To use our datasets in conjunction with **MongoDB** some preparations must be made, mainly substituting all **( )** with **[ ]** and inverting all coordinates (lat/long to long/lat). To make the process easier I've defined a simple function that takes care of it.

#Defining our converter function.

def town_converter(df):
    
    """
    Input : dataframe of towns.
    
    Output: dataframe with switched coordinates (lat/long -> long/lat) and substitutes all '(' with '['.
    
    """
    
    for i in range(len(df)):
        df['coords'].iloc[i] = '[' + str(eval(df['coords'].iloc[i])[1]) + ',' + str(eval(df['coords'].iloc[i])[0]) + ']'
        
    return df

#Converting our dataframe.

converted = town_converter(towns)

converted.head()

#Exporting our dataframe. The final 'i' marks it as an inverted dataframe, ready for MongoDB integration.

converted.to_csv('towns_1807_841i.csv', index=False)

# Ports: general data cleaning

Since this dataframe has had a lot of work behind it, it's already quite tidy and just needs a few tweaks. Column names will be translated into english.

ports = pd.read_csv('compartir/puertos.csv')

ports.info()

#Renaming the columns and adding a new one.

ports['photo'] = None
ports = ports.rename(columns = {'puerto': 'name', 'provincia': 'province', 'pueblo': 'municipality', 'altitud': 'altitude', 'desnivel': 'gradient', 'distancia': 'distance', 'pendiente': 'mountain_slope', 'coeficiente': 'technical_difficulty', 'coords': 'peak_coords'})

#Saving our dataframe:

ports.to_csv('compartir/puertos.csv', index=False)

#Before we save it for use on MongoDB we must perform the customary cleanup.

for i in range(len(ports)):
    ports['peak_coords'].iloc[i] = '[' + str(eval(ports['peak_coords'].iloc[i])[1]) + ',' + str(eval(ports['peak_coords'].iloc[i])[0]) + ']'

#Exporting our inverted dataframe.

ports.to_csv('compartir/puertos_i.csv', index=False)

# Conclusion and notes

Now our 3 dataframes are clean and ready for integration. Some additional features will be added at a later date, but the general format and layout will probably remain unchanged. 

This processing and use of functions will need to be structured into a pipeline for future data intake.

**<div align="right">Ironhack DA PT 2021</div>**
    
**<div align="right">Xavier Esteban</div>**
