==> 10. Adding ports to towns dataframe.ipynb <==
# Adding ports that can be accessed from each town

We need to find out which ports can be accessed from every destination. It can be done by iterating through our towns, routes and ports dataframes.

import pandas as pd

## Importing our dataframes

routes = pd.read_csv('routes_2207_636.csv')

routes.head(1)

routes.head(1)

ports = pd.read_csv('puertos_i.csv')

ports.head(1)

towns = pd.read_csv('towns_1907_201.csv')

towns.head(1)

## Adding ports to towns

While it isn't the most elegants of solutions, we will be using a simple loop to iterate through all towns, routes and ports to append the result to the corresponding row in the towns dataframe.

#Running our loop.

for i in range(len(towns)): #Iterating through each town.
    list_ports = [] #This list holds ports.
    list_routes = eval(towns['routes_ids'].iloc[i]) #Using eval to iterate through each port of a route as a list.
    for n in list_routes:
        port = routes[routes['ID'] == n]['mountain_passes_ids'].iloc[0] #Grabbing the port.
        list_ports.append(port) #Appending the list of ports to the town.
    towns['mountain_passes_ids'].iloc[i] = list_ports #Assinging the ports.

#Let's check the result.

towns.head(1)

Now we have the ports as a list of lists, with duplicated ports. Let's fix that using the unlimited power of loops.

for i in range(len(towns)): #Iterating through teach town.
    final_list = []
    for n in towns['mountain_passes_ids'].iloc[i]:
        for p in eval(n):
            if p not in final_list: #Not appending duplicates.
                final_list.append(p) #Appending the port.
    towns['mountain_passes_ids'].iloc[i] = final_list #Assigning the port list to the column.

#Checking the results.

towns.head()

#Exporting the results.

towns.to_csv('towns_2207_201.csv', index=False)

**<div align="right">Ironhack DA PT 2021</div>**
    
**<div align="right">Xavier Esteban</div>**
