==> 13. Cleaning up our ports dataframe, adding gpx file routes.ipynb <==
# Deleting ports with no routes

After cleaning up our routes database, there's some ports that are missing routes (temporarily). We must delete them and clean up all dataframes so thta we can work with them in **MongoDB**.

## Importing files and exploration

import pandas as pd

ports = pd.read_csv('puertos_i.csv')

ports.head()

routes = pd.read_csv('routes_2307_598.csv')

routes.head()

## Seeking out missing ports

Let's compare a list of all port IDs with the ones present in the *routes* dataframe.

#Creating a list of all ports.

ports_list = ports['ID'].tolist()

#Let's check how many of them are there.

len(ports_list)

#Creating a list of all individual ports present in our routes.

routes_ports = []

for i in range(len(routes)):
    for n in eval(routes['mountain_passes_ids'].iloc[i]):
        if n not in routes_ports:
            routes_ports.append(n)

#Let's check how many ports we have in our routes dataframe.

len(routes_ports)

#Now it's time to create a new list with the missing ones so that we can filter them out.

missing_ports = []

for i in ports_list:
    if i not in routes_ports:
        missing_ports.append(i)

len(missing_ports)

## Dropping our missing ports

#Since port ID matches the index position, we can simply drop them like so.

for i in missing_ports:
    ports = ports[ports['ID'] != i]

#Checking our results.

ports.head()

#Finally, reindexing and exporting our dataframe.

ports = ports.reset_index(drop=True)
ports.to_csv('ports_643.csv', index=False)

# Adding gpx file routes to each route

Now that our *gpx* files are hosted in our server (albeit with a temporary adress) we can finally add their web adress to the dataframe.

#Creating a new column to hold our adresses.

routes['gpx_link'] = None

#Since our adress is a known base url + the route ID and the .gpx format it's quite easy to populate the column.

base_url = 'http://www.on2wheels.es/gpx/' #Base url.
extension = '.gpx' #File extension.

for i in range(len(routes)):
    adress = base_url + str(routes['ID'].iloc[i]) + extension #Creating adress from base_url, ID and extension.
    routes['gpx_link'].iloc[i] = adress #Assigning it to the corresponding column.

#Everything seems in order.

routes.head()

#Exporting our dataframe.

routes.to_csv('routes_2707_598.csv', index=False)

**<div align="right">Ironhack DA PT 2021</div>**
    
**<div align="right">Xavier Esteban</div>**
