==> 15. Ranking our routes based on weather.ipynb <==
# Creating a climate index API

In order to know which destinations have (or will have) the best weather on a given date we will be using historical data from all of Spain's weather stations, as well as packaging that information into an API that will allow us to remotely perform queries.

Since all information from 291 weather stations are stored in individual *csv* files, let's import the first one and see how can we coax the necessary data out of it.

# Transforming historical data into a weather index

*Indices are particularly valuable because they allow the integrated effects of a range of climatic variables to be quantified, facilitating an interpretation and rating of climatic conditions at a destination. Another advantage of indices is that they enable the climate of tourism destinations to be objectively compared and are therefore a convenient and more conceptually sound means to assess possible impacts of climate change on the distribution of climatic resources worldwide.*

                  - An Inter-Comparison of the Holiday Climate Index (HCI) and the Tourism Climate Index (TCI) in Europe
                    https://www.mdpi.com/2073-4433/7/6/80/htm
                    
                    
                    
                    
Our aim is to condense all meteorological data (wind, temperature, hours of daylight, rain...) into a single numnber (an index) that can be used to easily compare one destination with another. The objective is to create a dataframe where every row is a different destination and each column holds the index for each week of the year.

We can't use the **TCI** as-is because we lack some of the data (humidity and cloud cover %), so we will have to make an index of our own. From now on we will call it simply the **CWI** (**C**ycling **W**eather **I**ndex).

## Creating our **CWI**

While it would be nice to have all weather data for every town in Spain, we have to make do with the available datasets.

Of the present variables in the official meteo datasets, the following variables are of use:


- Date (to group values per week).
- Average temperature.
- Rain (mm).
- Average wind speed.


The hours of sunlight aren't really meaningful because we're crafting this custom index to be used only on Spain, which has a single timezone and the latitude change isn't that great.


The original **TCI** roughly assigns the following weights to each parameter:


**Maximum temperature:** 40%

**Cloud cover:** 20%

**Precipitation:** 30%

**Wind:** 10%



While those percentages are quite good, this model was made with tourism in mind. Cycling has a few diferrences with tourism and other leisure activities that must be kept in mind when creating our **CWS**:

- Clouds are OK as long as it doesn't rain.
- Wind is very harmful.


A first approach to our index could be the following:

**Maximum temperature:** 40%

**Precipitation:** 40%

**Wind:** 20%

## Importing our data

The first step is importing our data. Every weather station's data is contained in a different *csv* file, we'll import the first one to see how data is structured.

import pandas as pd
import time
from pathlib import Path

#Importing our dataframe.

df = pd.read_csv('meteo_test.csv', sep=';')

df.info()

df.head()

#The precipitation values are split by time intervals, so we will just add them up to get a day's rain.

df['prec'] = None
df['prec'] = df['Prec1'] + df['Prec2'] + df['Prec3'] + df['Prec4']

## Using time series to get the average values

Since we want to calculate the **TCI** on a weekly basis it makes sense to compact our data with the same granularity. For this purpose we will be using time series.

#Converting 'FECHA' to Datetime format.

df["Fecha"] = df['Fecha'].apply(pd.to_datetime)

#Checking the result.

df.info()

Since our new datasets keep data from 2013 onwards will be using all of it. If we had data from some decades ago we would have to trim it. Now it's time to calculate the weekly averages.

#Using datetime to add the week number as a new column.

df['week'] = None
weekNumber = df['Fecha'].dt.week.tolist()
df['week'] = weekNumber

df.head()

#Creating a dataframe with the weekly mean values.

df_mean = df.groupby("week").mean()

## Assigning each value a rating 0-10

To create our index all values (temperature, rain and wind speed) will need to be rated on a given scale, from 0 to 10. For this we will be using the same ratings as the famous **TCI**.

#Assigning temperature values.

df_mean['tmax'] = None

for i in range(len(df_mean)):
    if df_mean['Tmax'].iloc[i] >= 39: #Checking the temperature, from high to low.
        df_mean['tmax'].iloc[i] = 0 #Assigning the score.
    elif df_mean['Tmax'].iloc[i] > 38:
        df_mean['tmax'].iloc[i] = 1
    elif df_mean['Tmax'].iloc[i] > 37:
        df_mean['tmax'].iloc[i] = 2
    elif df_mean['Tmax'].iloc[i] > 36:
        df_mean['tmax'].iloc[i] = 3
    elif df_mean['Tmax'].iloc[i] > 35:
        df_mean['tmax'].iloc[i] = 4
    elif df_mean['Tmax'].iloc[i] > 33:
        df_mean['tmax'].iloc[i] = 5
    elif df_mean['Tmax'].iloc[i] > 31:
        df_mean['tmax'].iloc[i] = 6
    elif df_mean['Tmax'].iloc[i] > 29:
        df_mean['tmax'].iloc[i] = 7
    elif df_mean['Tmax'].iloc[i] > 27:
        df_mean['tmax'].iloc[i] = 8
    elif df_mean['Tmax'].iloc[i] > 26:
        df_mean['tmax'].iloc[i] = 9
    elif df_mean['Tmax'].iloc[i] > 23:
        df_mean['tmax'].iloc[i] = 10
    elif df_mean['Tmax'].iloc[i] > 20:
        df_mean['tmax'].iloc[i] = 9
    elif df_mean['Tmax'].iloc[i] > 19:
        df_mean['tmax'].iloc[i] = 8
    elif df_mean['Tmax'].iloc[i] > 18:
        df_mean['tmax'].iloc[i] = 7
    elif df_mean['Tmax'].iloc[i] > 15:
        df_mean['tmax'].iloc[i] = 6
    elif df_mean['Tmax'].iloc[i] > 11:
        df_mean['tmax'].iloc[i] = 5
    elif df_mean['Tmax'].iloc[i] > 7:
        df_mean['tmax'].iloc[i] = 4
    elif df_mean['Tmax'].iloc[i] > 0:
        df_mean['tmax'].iloc[i] > 3
    elif df_mean['Tmax'].iloc[i] <= 3:
        df_mean['tmax'].iloc[i] = 0

#Assigning precipitation values.

df_mean['rain'] = None

for i in range(len(df_mean)):
    if df_mean['prec'].iloc[i] >= 40:
        df_mean['rain'].iloc[i] = -2
    elif df_mean['prec'].iloc[i] >= 25:
        df_mean['rain'].iloc[i] = -1
    elif df_mean['prec'].iloc[i] > 12:
        df_mean['rain'].iloc[i] = 0
    elif df_mean['prec'].iloc[i] > 10:
        df_mean['rain'].iloc[i] = 1
    elif df_mean['prec'].iloc[i] > 9:
        df_mean['rain'].iloc[i] = 2
    elif df_mean['prec'].iloc[i] > 8:
        df_mean['rain'].iloc[i] = 3
    elif df_mean['prec'].iloc[i] > 7:
        df_mean['rain'].iloc[i] = 4
    elif df_mean['prec'].iloc[i] > 6:
        df_mean['rain'].iloc[i] = 5
    elif df_mean['prec'].iloc[i] > 5:
        df_mean['rain'].iloc[i] = 6
    elif df_mean['prec'].iloc[i] > 4:
        df_mean['rain'].iloc[i] = 7
    elif df_mean['prec'].iloc[i] > 3:
        df_mean['rain'].iloc[i] = 8
    elif df_mean['prec'].iloc[i] > 0:
        df_mean['rain'].iloc[i] = 9
    elif df_mean['prec'].iloc[i] == 0.0:
        df_mean['rain'].iloc[i] = 10

#Assigning wind values.

df_mean['wind'] = None

for i in range(len(df_mean)):
    if df_mean['Vmax'].iloc[i] > 70:
        df_mean['wind'].iloc[i] = -10
    elif df_mean['Vmax'].iloc[i] > 65:
        df_mean['wind'].iloc[i] = -8
    elif df_mean['Vmax'].iloc[i] > 60:
        df_mean['wind'].iloc[i] = -6
    elif df_mean['Vmax'].iloc[i] > 55:
        df_mean['wind'].iloc[i] = -4
    elif df_mean['Vmax'].iloc[i] > 50:
        df_mean['wind'].iloc[i] = -2
    elif df_mean['Vmax'].iloc[i] > 40:
        df_mean['wind'].iloc[i] = 0
    elif df_mean['Vmax'].iloc[i] > 35:
        df_mean['wind'].iloc[i] = 2
    elif df_mean['Vmax'].iloc[i] > 30:
        df_mean['wind'].iloc[i] = 6
    elif df_mean['Vmax'].iloc[i] > 25:
        df_mean['wind'].iloc[i] = 7
    elif df_mean['Vmax'].iloc[i] > 20:
        df_mean['wind'].iloc[i] = 8
    elif df_mean['Vmax'].iloc[i] > 10:
        df_mean['wind'].iloc[i] = 9
    elif df_mean['Vmax'].iloc[i] < 10:
        df_mean['wind'].iloc[i] = 10

df_mean.head()

## Assigning a CWI score to each week

Having all relative scores for any given week (max temperature, rain and wind speed) we can proceed to assign a score.

#Creating a new column by multiplying our individual scores by the corresponding %.

df_mean['CWI'] = df_mean['tmax']*0.4 + df_mean['rain']*0.4 + df_mean['wind']*0.2

#Our final dataframe needs only the last column (score).

df_score = pd.DataFrame(df_mean['CWI'])

df_score.head()

#Transposing the dataframe to condense all information in a single row.

df_t = df_score.T

#Reseting the index.

df_t.reset_index(drop=True, inplace=True)

#Finally, creating new columns for the weather station name and coordinates.

df_t['name'] = None
df_t['coords'] = None

df_t.head()

# Applying the scoring system to every weather station

Now that we have all the necessary steps to transform a dataframe into our desired structure it's time to create a function (or a loop) that can automatically perform this operation on all *csv* containing meteo data. To access each folder we will be using **Path**.

This will be done in three steps, with three separate functions/loops:

1. Iterate through every *csv* file in a folder.
2. Transform each *csv* file with the previous methods and return a dicionary.
3. Create a master dataframe using all generated dictionaries.

## Creating our parser

Let's begin by defining the parameters that our function must meet:

1. Perform all necessary transformations on a given *csv* file.
2. Return a dictionary where the first key is the station name, the second one the coords, and the next 53 correspond to a week with their respective values reflecting that week's **CWI**.

df.head(1)

#Defining our function.

def score_creator(file):
    df = pd.read_csv(file, sep=';') #Opening the file.

    df['Id'] = df['Id'].astype(str) #Typecasting as a str, otherwise the zeros on the left will be lost.
    
    df['prec'] = None
    df['prec'] = df['Prec1'] + df['Prec2'] + df['Prec3'] + df['Prec4']

    df["Fecha"] = df['Fecha'].apply(pd.to_datetime) #Converting from object to Datetime.
     
    df['Tmax'].fillna(df['Tmax'].mean(), inplace=True) #Replacing NaNs.
    df['prec'].fillna(df['prec'].mean(), inplace=True)
    df['Vmax'].fillna(df['Vmax'].mean(), inplace=True)
    
    newname = df['Id'].iloc[1]
    
    df['week'] = None
    weekNumber = df['Fecha'].dt.week.tolist()
    df['week'] = weekNumber #Grouping by week.
    
    df_mean = df.groupby("week").mean() #Calculating the weekly average values.
    
    df_mean['tmax'] = None

    for i in range(len(df_mean)):
        try:
            if df_mean['Tmax'].iloc[i] >= 39: #Checking the temperature, from high to low.
                df_mean['tmax'].iloc[i] = 0 #Assigning the score.
            elif df_mean['Tmax'].iloc[i] > 38:
                df_mean['tmax'].iloc[i] = 1
            elif df_mean['Tmax'].iloc[i] > 37:
                df_mean['tmax'].iloc[i] = 2
            elif df_mean['Tmax'].iloc[i] > 36:
                df_mean['tmax'].iloc[i] = 3
            elif df_mean['Tmax'].iloc[i] > 35:
                df_mean['tmax'].iloc[i] = 4
            elif df_mean['Tmax'].iloc[i] > 33:
                df_mean['tmax'].iloc[i] = 5
            elif df_mean['Tmax'].iloc[i] > 31:
                df_mean['tmax'].iloc[i] = 6
            elif df_mean['Tmax'].iloc[i] > 29:
                df_mean['tmax'].iloc[i] = 7
            elif df_mean['Tmax'].iloc[i] > 27:
                df_mean['tmax'].iloc[i] = 8
            elif df_mean['Tmax'].iloc[i] > 26:
                df_mean['tmax'].iloc[i] = 9
            elif df_mean['Tmax'].iloc[i] > 23:
                df_mean['tmax'].iloc[i] = 10
            elif df_mean['Tmax'].iloc[i] > 20:
                df_mean['tmax'].iloc[i] = 9
            elif df_mean['Tmax'].iloc[i] > 19:
                df_mean['tmax'].iloc[i] = 8
            elif df_mean['Tmax'].iloc[i] > 18:
                df_mean['tmax'].iloc[i] = 7
            elif df_mean['Tmax'].iloc[i] > 15:
                df_mean['tmax'].iloc[i] = 6
            elif df_mean['Tmax'].iloc[i] > 11:
                df_mean['tmax'].iloc[i] = 5
            elif df_mean['Tmax'].iloc[i] > 7:
                df_mean['tmax'].iloc[i] = 4
            elif df_mean['Tmax'].iloc[i] > 0:
                df_mean['tmax'].iloc[i] > 3
            elif df_mean['Tmax'].iloc[i] <= 3:
                df_mean['tmax'].iloc[i] = 0
        except:
            df_mean['tmax'].iloc[i] = df_mean['tmax'].mean() #Filling missing values with the mean.
            
    df_mean['rain'] = None
    for i in range(len(df_mean)):
        try:
            if df_mean['prec'].iloc[i] >= 40:
                df_mean['rain'].iloc[i] = -2
            elif df_mean['prec'].iloc[i] >= 25:
                df_mean['rain'].iloc[i] = -1
            elif df_mean['prec'].iloc[i] > 12:
                df_mean['rain'].iloc[i] = 0
            elif df_mean['prec'].iloc[i] > 10:
                df_mean['rain'].iloc[i] = 1
            elif df_mean['prec'].iloc[i] > 9:
                df_mean['rain'].iloc[i] = 2
            elif df_mean['prec'].iloc[i] > 8:
                df_mean['rain'].iloc[i] = 3
            elif df_mean['prec'].iloc[i] > 7:
                df_mean['rain'].iloc[i] = 4
            elif df_mean['prec'].iloc[i] > 6:
                df_mean['rain'].iloc[i] = 5
            elif df_mean['prec'].iloc[i] > 5:
                df_mean['rain'].iloc[i] = 6
            elif df_mean['prec'].iloc[i] > 4:
                df_mean['rain'].iloc[i] = 7
            elif df_mean['prec'].iloc[i] > 3:
                df_mean['rain'].iloc[i] = 8
            elif df_mean['prec'].iloc[i] > 0:
                df_mean['rain'].iloc[i] = 9
            elif df_mean['prec'].iloc[i] == 0.0:
                df_mean['rain'].iloc[i] = 10
        except:
            df_mean['rain'].iloc[i] = df_mean['rain'].mean() #Filling missing values with the mean.
            
    df_mean['wind'] = None
    for i in range(len(df_mean)):
        try:
            if df_mean['Vmax'].iloc[i] > 70:
                df_mean['wind'].iloc[i] = -10
            elif df_mean['Vmax'].iloc[i] > 65:
                df_mean['wind'].iloc[i] = -8
            elif df_mean['Vmax'].iloc[i] > 60:
                df_mean['wind'].iloc[i] = -6
            elif df_mean['Vmax'].iloc[i] > 55:
                df_mean['wind'].iloc[i] = -4
            elif df_mean['Vmax'].iloc[i] > 50:
                df_mean['wind'].iloc[i] = -2
            elif df_mean['Vmax'].iloc[i] > 40:
                df_mean['wind'].iloc[i] = 0
            elif df_mean['Vmax'].iloc[i] > 35:
                df_mean['wind'].iloc[i] = 2
            elif df_mean['Vmax'].iloc[i] > 30:
                df_mean['wind'].iloc[i] = 6
            elif df_mean['Vmax'].iloc[i] > 25:
                df_mean['wind'].iloc[i] = 7
            elif df_mean['Vmax'].iloc[i] > 20:
                df_mean['wind'].iloc[i] = 8
            elif df_mean['Vmax'].iloc[i] > 10:
                df_mean['wind'].iloc[i] = 9
            elif df_mean['Vmax'].iloc[i] < 10:
                df_mean['wind'].iloc[i] = 10
        except:
            df_mean['wind'].iloc[i] = df_mean['wind'].mean() #Filling missing values with the mean. 
            
    df_mean['CWI'] = df_mean['tmax']*0.4 + df_mean['rain']*0.4 + df_mean['wind']*0.2 #Calculating the score.
    
    df_score = pd.DataFrame(df_mean['CWI']) #Keeping just the score.
    
    df_t = df_score.T #Transposing the dataframe.
    
    df_t['name'] = None
    df_t['coords'] = None
    
    df_t.reset_index(drop=True, inplace=True) #Resetting index.
    
    c = df_t.columns.tolist() #Creating a list of columns.
    r = df_t.iloc[0].tolist() #Creating a list of row values.
    
    keys = c #Creating a dictionary from our two lists.
    values = r
    dictionary = dict(zip(keys, values))
    
    dictionary['name'] = newname
    
    return dictionary #Returning our dictionary.

output = score_creator('meteo_test.csv')

#Testing our function.

output

## Creating the component that iterates through all csv files and running it

#This loop will iterate through all files in our designated folder and apply the yet to be defined function to them.

start = time.time() #Starting a timer.

dict_list = [] #We will be appending our output dictionaries here.

directory = 'csv' #The folder containing our csv files.
 
files = Path(directory).glob('*') #Grabbing all files.
for file in files:
    dict_list.append(score_creator(file)) #Applying our score function.
    
stop = time.time() #Stopping our timer.
duration = (stop - start) / 60
print('Minutes:', duration) #Returning the elapsed minutes.

test = pd.DataFrame(dict_list) #Creating the dataframe from the list of dictionaries.

We can see that many rows (weather stations) are missing or have missing weeks, that's because they have faulty data or the recording simply doesn't fill the last 5 years. The rows with missing data can be safely dropped.

test.info()

#Dropping the rows with missing data.

meteo = test.dropna(thresh=52)

#Finally, saving our resulting dataframe.

meteo.to_csv('meteo.csv', index=False)

# Adding the weather station coordinates

Now that we have our complete dataframe it's to assign every weather station its geographical coordinates. This can be done quite easily by using *ListadoEstaciones*, a *csv* file containing all identifiers and corresponding coordinates.

listado = pd.read_csv('ListadoEstaciones.csv', sep=';')

listado.head()

Latitude and longitude aren't expressed in degrees but in degrees, minutes and seconds. Let's define two functions to convert them.

def latDD(x):
    '''
    Input: latitude in DDMMSS.
    
    Output: latitude in degrees.
    
    '''
    D = int(x[0:2])
    M = int(x[2:4])
    S = float(x[4:6])
    DD = D + float(M)/60 + float(S)/3600
    DD = round(DD, 5) #Setting a limit to 5 decimals.
    return DD

def longDD(x):
    '''
    Input: longitude in DDMMSS.
    
    Output: longitude in degrees.
    
    '''    
    D = int(x[0:2])
    M = int(x[2:4])
    S = float(x[4:6])
    DD = D + float(M)/60 + float(S)/3600
    if 'W' in x:
        DD = (0 - DD)
    DD = round(DD, 5)
    return DD

#Testing the first function.

latDD(listado['lat'].iloc[0])

#Testing the second one.

longDD(listado['long'].iloc[0])

#Now let's apply the functions to the dataframe.

listado['lat'] = listado['lat'].map(lambda x: latDD(x))
listado['long'] = listado['long'].map(lambda x: longDD(x))

listado.head()

#Now we can finally add the coordinates to the original meteo dataframe.

for i in range(len(listado)):
    for n in range(len(meteo)):
        if listado['ID'].iloc[i] == meteo['name'].iloc[n]:
            meteo['coords'].iloc[n] = '('+ str(listado['lat'].iloc[i]) + ',' + str(listado['long'].iloc[i]) + ')'

#Checking the results.

meteo.head()

#Saving our dataframe.

meteo.to_csv('meteo.csv', index=False)

**<div align="right">Ironhack DA PT 2021</div>**
    
**<div align="right">Xavier Esteban</div>**
